Migrated styleSelector state into overview
